Fix legacy PXE booting when running dhcp and PXE on same server.
When running both DHCP and PXE on the same server we need to direct the
PXE clients over to port 4011 for the boot file. To do this we don't set
the bootfile in the dhcp offer with the client ip. This will direct the
client to send another request to port 4011 for the bootfile.