Fix bug with RTC timezone on kickstarts.
After kicking a node, RTC in local TZ was set to yes.

```
 timedatectl status
      Local time: Tue 2020-01-21 13:56:14 EST
  Universal time: Tue 2020-01-21 18:56:14 UTC
        RTC time: Tue 2020-01-21 13:56:14
       Time zone: America/New_York (EST, -0500)
     NTP enabled: yes
NTP synchronized: yes
 RTC in local TZ: yes

Warning: The system is configured to read the RTC time in the local time zone.
         This mode can not be fully supported. It will create various problems
         with time zone changes and daylight saving time adjustments. The RTC
         time is never updated, it relies on external facilities to maintain it.
         If at all possible, use RTC in UTC by calling
         'timedatectl set-local-rtc 0'.
```

The fix for this was to specify the `--utc` option to the timezone anaconda
command. More info here: https://pykickstart.readthedocs.io/en/latest/kickstart-docs.html#timezone