When discovering via SNMP ports are off by one.
When testing this on a few Dell switches we have, it appears querying
the MAC table via SNMP returns different port numbers that if querying
via the API. It looks like the port numbers via SNMP are off by one.

TODO: Figure out a better way of handling this but for now it works for
all our switches.