Add endpoint to unprovision host.
One use case for this is to support legacy booting. In the post of a
kickstart we send an HTTP POST request (via curl) to this endpoint which
turns off provisioning for the host. This prevents a kickstart loop as
after the node reboots it will not be sent a PXE boot response.

When uefi booting this is not necessary as the grub uefi will set the
first boot device to be the hard drive after completing the kickstart
installation.